Splitting up files in messages
